+++ crates/amaru-ledger/src/governance/ratification/proposals_forest.rs
@@ -731,11 +731,12 @@ mod tests {
     };
 
     use amaru_kernel::{
-        Anchor, Bytes, ConstitutionalCommitteeUpdate, Epoch, GovernanceAction, Hash, KeyValuePairs, Lovelace,
-        MaxString128, OrphanProposal, PREPROD_DEFAULT_PROTOCOL_PARAMETERS, PROTOCOL_VERSION_10, Proposal, ProposalEnum,
-        ProposalId, ProposalPointer, ProposalsRootsRc, ProtocolParameters, RationalNumber, Slot, TransactionPointer,
-        any_constitution, any_constitutional_committee_update, any_gov_action, any_proposal_enum, any_proposal_id,
-        any_proposal_pointer, any_protocol_params_update, any_protocol_version, any_reward_account,
+        Anchor, ConstitutionalCommitteeUpdate, Epoch, GovernanceAction, Hash, KeyValuePairs, Lovelace, MaxString128,
+        Network, OrphanProposal, PREPROD_DEFAULT_PROTOCOL_PARAMETERS, PROTOCOL_VERSION_10, Proposal, ProposalEnum,
+        ProposalId, ProposalPointer, ProposalsRootsRc, ProtocolParameters, RationalNumber, RewardAccount, Slot,
+        StakeCredential, TransactionPointer, any_constitution, any_constitutional_committee_update, any_gov_action,
+        any_proposal_enum, any_proposal_id, any_proposal_pointer, any_protocol_params_update, any_protocol_version,
+        any_reward_account,
     };
     use proptest::{collection, prelude::*, test_runner::RngSeed};
 
@@ -1361,7 +1362,7 @@ mod tests {
     fn make_proposal(parent: Option<ProposalId>) -> Proposal {
         Proposal {
             deposit: 0,
-            reward_account: Bytes::default(),
+            reward_account: RewardAccount::new(Network::Testnet, StakeCredential::KeyHash(Hash::new([0; 28]))),
             gov_action: GovernanceAction::HardForkInitiation(parent, PROTOCOL_VERSION_10),
             anchor: Anchor {
                 url: MaxString128::from_str("https://example.com").unwrap(),